Pking moj/helm deploy (#84)
## Description of change

Provide the ability to deploy the application to multiple environments
using Helm as the kubernetes tool to assist with multple envs. Also
updated the gitactions to add in protection/checks when deploying

## Link to relevant ticket

n/a

## Notes for reviewer

Dev is still using the old method until we are happy with Helm. Prod can
only be deployed to main with an approval from a Senior Engineer (Ivan
and Paul currently)

## Screenshots of changes (if applicable)

### Before changes:

### After changes:

## How to manually test the feature

---------

Co-authored-by: David Henry <david.henry@digital.justice.gov.uk>