Crm457 237/dont presave (#122)
## Description of change
The system currently pre-creates the records before any data is filled
in as this was easier to manage initially. This has the negative side
effect that is it easy to generate bad data in the sytsem that does not
make a heap of sense.

We should only save when the user has clicked a save button.

## Link to relevant ticket
https://dsdmoj.atlassian.net/browse/CRM457-237

## NOTES to reviewer
I have used a fake ID to allow pre-created records to be accessed via
the edit and update screens. There is an argument that we should instead
use new/create in these instances, however that is a much larger
reworking of the code so at this stage I have chosen not to do that.
Happy to discuss