[CRM457-213] Solicitor declaration step (#147)
## What
This is the page on the form that allows a user to confirm what they
have submitted is correct.

## Link to relevant ticket
[SOLICITOR DECLARATION: Build "Confirm the following"
page](https://dsdmoj.atlassian.net/browse/CRM457-213)

## Notes for reviewer
After submission of the claim, status is set to 'completed'. We do need
to confirm what to show when user clicks on the 'SUBMITTED' claim in the
Main Claim list page. [ref.Screenshot below]. I left a placeholder with
a TODO tag for us to change the URL later.
![Screenshot 2023-07-13 at 12 26
38](https://github.com/ministryofjustice/laa-claim-non-standard-magistrate-fee-backend/assets/90189839/6e16717c-9379-48e9-9692-a10e4a11e1cd)


## Screenshot
![Screenshot 2023-07-13 at 12 22
45](https://github.com/ministryofjustice/laa-claim-non-standard-magistrate-fee-backend/assets/90189839/2226afcc-790e-4a45-91f8-ee2fe243e13e)