Remove defendants from shared example, only included where needed
At least a main defendant is needed to show the check answers page. Therefore when showing the page or clicking to navigate to the page (in Other Info) we need to have this populated.

I could instead create/use a class to store the  attributes of a defendant and then build instantiations of that from the claim - but this feels overkill