CRM457-1017: Additional cost screens (#475)
## Description of change
Add the additional cost screens

## Link to relevant ticket
https://dsdmoj.atlassian.net/browse/CRM457-1017

## Notes for reviewer
- The 'add another' pattern of NSM is quite unwieldy, requiring repeated
checks of the ID against a 'fake' one of 0000-0000-000000. The built-in
Rails `resources` paradigm basically does what we wanted, so I chose
that for this.
- That did mean making a couple of changes, mainly changing `:id` to
`:application_id` in the URL structure, which had a couple of knock-on
effects, but nothing significant - it turns out a lot of the code that
set this ID explicitly was covering up some stuff that Rails inferred
automatically
- The wiring up of these screens to the rest of the flow is incomplete,
because the screens these connect to don't exist yet.
- Speaking of which, further changes are brewing. I'll keep an eye on
the Slack for the next couple of days to see where they land. At least
one of these screens might be going away.

## Screenshots of changes (if applicable)

### Before changes:

### After changes:

## How to manually test the feature