CRM457-1040: Fix bug in navigation stack logic (#512)
## Description of change

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1040)

## Notes for reviewer
Previously if you completed steps A and B then went back and edited step
A, step B would be marked as 'not started'. Which would be extra
problematic if step B came before step A in the task list.

As we have our own logic for determining where changes in A invalidate
step B, the simplest thing is to simply remove the 'prune' mechanism
that's causing the app to forget about previous interactions.