CRM457-979: Add navigation flow from CYA page
The prototype confirms, in most instances, that flow
from the CYA page should return to the CYA page once
a cards page (or group of pagess) is complete. It should
not return to the tasklist page.

The prototype may be defective as it does not do this
in all instances (Crown court hearing details returns
to tasklist).

The primary quote card also needs special handling
as the primary quote summary page functions ooutside
the nortmal decision tree flow. Its save and continue
button desitination is hard coded.