CRM457-979: Move `actions` header resposibility to cards and add row actions capability
The Change action links in the card "header" are the responsibility
of the individual card objects rather than the report object.

One card, Application details, should not have this link on
the header but on the Ufn row instead, according to designs
and, by moving the responsibility to the card object, this
becomes easy.

In addition, the govuk_summary_list accepts an ` actions:` key
on the individual rows so we use this to add a link to a specifc
row (UFN).