Add test suite into main branch pipeline
We should run the test suite on merge to main
again. This reduces risk of breakages being
released. Additionally it renders the need
to rebase branches before merge less important.

We have also decided to only build after tests
pass. This is expensive in time and may be changed
in the future.