Add clamav install script and amend clamby dependent test
This adds a tested install script that also performs the
necessary configuration for having clamav work locally on
mac os.

In addition the tests for the file uploader, which were failing
locally have been fixed to stub out the actual clamav calls
since we are not using an actual "infected" file (and the tests
are more to excercise the ENV var and host environement config
impact on scanning in any event).