Add workaround to missing translation issue and fix defendant card (#620)
## Description of change
The 'check answers' screen for NSM was showing 'incomplete' for
defendants because the card expected a "full_name" field and the model
only had "first_name" and "last_name" (and the test strategy is so
heavily dependent on unit tests that mock everything they interact with
that this mismatch wasn't shown up). This fixes the card.

For bonus points, it adds a more thorough workaround and documentation
for the bug where the check answers screen explodes in dev because
things that are not translation keys are being passed round as
translation keys.