Do not inline sidekiq for bin/dev
default to running sidekiq inline for `rails s`
but not when using `bin/dev`. This is so `bin/dev` behaves
as a hosted environments sidekiq worker as expected and
stats are updated.

Without this the job will be processed inline with the side
effect that stats on the webui are not updated.