CRM457-1193: Ensure provider accepts confirmations
This removes the columns entirely and uses virtual attributesfor
confirmations rather than resetting these. It should have been done
like this to start with as providers should always be expected to
confirm the latest info on the page in any event, but especially
when the application has changed status (to sent_back) and needs
to be resubmitted.