Stop treating work items as invalid if uplift allowed but not applied (#741)
## Description of change
If you save a work item that is eligible for uplift, but don't apply an
uplift, it gets saved to the database with a zero uplift. But then the
dodgy form logic decides to say that that uplift _has_ been applied - an
uplift of zero. And that's now invalid.

This PR changes the form logic so that if a work item is eligible for
uplift and what has been saved is zero, it now correctly interprets that
as uplift _not_ having been applied, which is a valid state.