Only validate next hreaing date when nil or changed
This is too allow next hearing dates to be entered as
current/future dates once, but to ensure that if that
date becomes a past date after that it will not be considered
invalid. This is to prevent invalidating the step after it has
been completed, in particular following RFI requests for information.