CRM457-1401: Stop preselecting radio button (#770)
## Description of change
We infer the alternate contact details value from the presence of
conditionally revealed details when the page is first loaded. But when
the page has not previously been saved, the absence of contact details
should cause the radio to be unselected. This PR does that.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1401)