Remove meta_request (#884)
## Description of change
For reasons not clear, this gem is interfering with Rails form objects,
meaning that in dev mode on local, we're getting SystemStackErrors when
trying to view pages that pass the form object to the multifile_upload
partial. As this gem is only a development helper added in passing in
another PR, I think we can safely remove it.