Ensure save buttons disabled during upload without flicker
Previous mechanism resulted in flicker because it was re-enabling
and disabling on multifile uploads when one of the uploads
completed but others were still in progress.

This mechanism overides the MOJFrontend uploadFiles (not the plural)
method and leverages promises and async functions to wait for ALL
uploads to complete before re-enabling.

It also changes the behaviour to NOT hide/remove failed uploads but rather
display them with a warning icon. This is because in such instances
the top level feedback warning may be overwritten by successful uploads
in the same drag and drop batch leaving the user unaware their upload for
one or more of their files failed.