CRM457: Fix bug in deactivated office check (#914)
## Description of change
Correct behaviour: If all of a provider's office codes are both enrolled
and deactivated, the provider should only see the 'deactivated office'
error screen on attempting to log in.

What's actually happening: This works if a provider has multiple office
codes, all enrolled and deactivated. But if the provider has only one
office code, and that office code is enrolled and deactivated, the
provider can log in as normal.