Move id from header row to item button
Handles accessibility failings for links in table rows.

Alternative link text is contructed by the browser for any links
in a table row. For example, "Item 1 Car mileage", "Duplicate Item 1 Car mileage"
or "Delete Item 1 Car mileage". Previously the "Item"
text was not being found as the id `itemTitle` was on the header cell `th`
containg the form with a button with the text 'Item'. This may be a bug
with the accessibility tooling or browser itself (firefox), but nonetheless
we can fix this by moving the id to the button of the form.