CRM457-1905b: Remove redundant code (#1093)
## Description of change
We don't save confirmation fields to the DB, so in production we never
actually send them to the app store. But the tests act like we do. Since
we don't, and we don't need to, we can clear out a bunch of redundant
code that misleadingly makes it look like we persist this stuff and send
it to the app store.


[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1905)