CRM457-2101: Move clam to separate deployment/pods (#1231)
## Description of change
- Instead of having clamav in a container on every app pod, have it
running as its own pod in its own deployment accessed via its own
service
- This allows clamav to have its own liveness probe, and not have to be
restarted every time new app pods are rolled out
- Switch to using the latest image
- Only have one clam service on dev, not one per branch, reusing the
same mechanism employed for the custom error service
[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2101)

## Note to reviewer
When this is merged, dev main will be responsible for the dev clam
service, and branches will simply use the service. However, for the
purposes of this getting clamby working for this PR we need the branch
to run the service. Hence in the second commit I've hacked the config to
have the branch deploy the service, and added 3 'TODO' statements to
revert it on merge to giving responsibility to main.