CRM457-2106: Backfill timestamps (#1258)
## Description of change
Now that we're tracking when a submission was originally submitted (to
filter on by search) we need to backfill all of these timestamps for
previous submissions. I got the from the timestamp when the production
app store submission record was originally created, and dumped them into
CSVs. We iterate through and update. There's only a few thousand so it
only takes a few seconds, so I'm not super concerned about efficiency
here.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2106)