CRM457-2162: Tweak task list behaviour (#1251)
## Description of change
Override NSM task list to change labelling behaviour. Previously, if
you'd completed a sectionB but then amended a previous section A,
removing B from the navigation stack, B would be marked as 'Not started'
even though we still had all of the previously entered data for B. This
made it look like we had thrown data away. Now we mark 'B' as in
progress (by detecting that, despite not being in the stack, it is
complete).

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2162)