CRM457-2340: Stop looking in local DB unnecessarily (#1354)
## Description of change
Only refer to local DB post-submission when we really need to. This
means:
- For claims, we now don't load the local value at all if we're just
viewing a record
- During an RFI loop we only use the local DB to store the new
information provided by a provider until we're ready to push it back to
the app store
- We can even handle a claim RFI loop for a claim that doesn't exist at
all in the local DB (although only by creating a stub record so that the
local FI record foreign key doesn't break, which isn't ideal but
probably OK given it won't ever happen in production)
- PriorAuthorityApplications are a bit more complicated as to do an RFI
loop, even if it's just for further info, we still have to go via the
check answers screen and that gets a bit hairy. And obviously for
"corrections needed" RFIs we have to keep using the local DB. So I've
left PA RFIs looking in the local DB for now, but just viewing a
submitted record is already entirely avoiding the local DB.
- As a bonus, I've removed about 30 identical 1-line
`decision_tree_class` methods from controllers that all inherit from the
same base class

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2340)