CRM457-2217: Simplify event builder (#1357)
## Description of change
"Events" are getting simpler. We don't need them to list what sections
have changed. All we need is for `corrected_info` to be truthy if the
provider responded to a request for corrections, and falsey if it was
just a request for further info. This allows us to simplify the code
somewhat.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2217)