CRM457-2273: If validation fails, calculation should not explode (#1367)
## Description of change
If the form fails validation, invalid number attributes may be held as
strings (so we can accurately pre-populate the relevant field with the
exact wrong input the user provided). But we will still be calling the
calculator with the form values we have. To avoid it exploding, we
`to_i` the values, so that if they are strings they will be treated as
zeroes for the purposes of the calculation.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2273)