Re-enable skipped Chrome tests & shift back to 100% coverage
Thanks to
https://mojdt.slack.com/archives/C014YJ90G48/p1745834811181699, we
were made aware that there is a genuine issue to track against
Capybara.

So in lieu of an actual fix, there is a package capybara-lockstep
which correctly waits for components to be available before attempting
tests. This fixes whatever the random issue with Chrome was and lets
us go back to 100% coverage.